# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Burnout Monitor API
  description: API for employee burnout detection and peak performance identification
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /notifications/sms/config:
    get:
      tags: [Notifications]
      summary: Get SMS notification preferences
      responses:
        '200':
          description: SMS configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  smsEnabled:
                    type: boolean
                  phoneNumber:
                    type: string
                  onBurnout:
                    type: boolean
                  onOpportunity:
                    type: boolean

    post:
      tags: [Notifications]
      summary: Update SMS notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                smsEnabled:
                  type: boolean
                phoneNumber:
                  type: string
                onBurnout:
                  type: boolean
                onOpportunity:
                  type: boolean
      responses:
        '200':
          description: Configuration updated
  /notifications/sms/logs:
    get:
      tags: [Notifications]
      summary: Get SMS send history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: SMS logs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    phoneNumber:
                      type: string
                    messageBody:
                      type: string
                    status:
                      type: string
                    sentAt:
                      type: string
                      format: date-time

  # ==========================================
  # Demo Mode
  # ==========================================
...